pyproject.toml
[project]
name = "fitbit-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.76.0",
    "black>=26.1.0",
    "duckdb>=1.4.3",
    "duckdb-engine>=0.17.0",
    "ipykernel>=7.1.0",
    "isort>=7.0.0",
    "jupyter>=1.1.1",
    "kagglehub>=0.4.1",
    "langchain>=1.2.6",
    "langchain-anthropic>=1.3.1",
    "langchain-chroma>=1.1.0",
    "langchain-community>=0.4.1",
    "langchain-core>=1.2.7",
    "langchain-ollama>=1.0.1",
    "langchain-openai>=1.1.7",
    "langchain-text-splitters>=1.1.0",
    "langgraph>=1.0.6",
    "notebook>=7.5.2",
    "pip>=25.3",
    "pytest>=9.0.2",
    "python-dotenv>=1.2.1",
    "rich>=14.2.0",
    "streamlit>=1.53.0",
]
